move credential image related logic to MintCredentialContext
