+++ src/components/[guild]/Requirements/components/GuildCheckout/MintCredentialContext.tsx
@@ -2,12 +2,14 @@ import useGuild from "components/[guild]/hooks/useGuild"
 import useGuildPermission from "components/[guild]/hooks/useGuildPermission"
 import useIsMember from "components/[guild]/hooks/useIsMember"
 import useUser from "components/[guild]/hooks/useUser"
+import useLocalStorage from "hooks/useLocalStorage"
 import {
   createContext,
   Dispatch,
   PropsWithChildren,
   SetStateAction,
   useContext,
+  useEffect,
   useState,
 } from "react"
 import useSWRImmutable from "swr/immutable"
@@ -23,10 +25,33 @@ export enum GuildAction {
   IS_ADMIN,
 }
 
+const MIN_IMAGE_WH = 512
+
+type ImageWH = { width: number; height: number }
+
+const getImageWidthAndHeight = async (
+  _: string,
+  imageUrl: string
+): Promise<ImageWH> =>
+  new Promise((resolve) => {
+    const img = new Image()
+    img.onload = () => {
+      const { width, height } = img
+      resolve({
+        width,
+        height,
+      })
+    }
+    img.src = imageUrl
+  })
+
 const MintCredentialContext = createContext<
   {
     credentialType: GuildAction
     credentialImage: string
+    isImageValidating: boolean
+    isInvalidImage?: boolean
+    isTooSmallImage?: boolean
     error: string
     mintedTokenId?: number
     setMintedTokenId: Dispatch<SetStateAction<number>>
@@ -38,7 +63,9 @@ const MintCredentialProviderComponent = ({
 }: PropsWithChildren<unknown>): JSX.Element => {
   const guildCheckoutContext = useGuildCheckoutContext()
 
-  const { id } = useGuild()
+  const { id, imageUrl } = useGuild()
+  const isInvalidImage = !imageUrl || imageUrl?.startsWith("/guildLogos")
+
   const isMember = useIsMember()
   const { isOwner, isAdmin } = useGuildPermission()
   const { isSuperAdmin } = useUser()
@@ -51,8 +78,36 @@ const MintCredentialProviderComponent = ({
     ? GuildAction.JOINED_GUILD
     : null
 
-  const { data: credentialImage, error } = useSWRImmutable(
-    id && typeof credentialType === "number"
+  const [imageWHFromLocalstorage, setImageWHFromLocalstorage] =
+    useLocalStorage<ImageWH>(imageUrl ? `guildImageWH:${imageUrl}` : null, undefined)
+
+  const { data, isValidating } = useSWRImmutable(
+    !isInvalidImage && !imageWHFromLocalstorage
+      ? ["imageWidthAndHeight", imageUrl]
+      : null,
+    getImageWidthAndHeight
+  )
+
+  useEffect(() => {
+    if (!data || imageWHFromLocalstorage) return
+    setImageWHFromLocalstorage(data)
+  }, [data, imageWHFromLocalstorage])
+
+  const isTooSmallImage = imageWHFromLocalstorage
+    ? imageWHFromLocalstorage.width < MIN_IMAGE_WH ||
+      imageWHFromLocalstorage.height < MIN_IMAGE_WH
+    : !isValidating &&
+      data &&
+      (data.width < MIN_IMAGE_WH || data.height < MIN_IMAGE_WH)
+
+  const shouldFetchImage =
+    id && typeof credentialType === "number" && !isInvalidImage && !isTooSmallImage
+  const {
+    data: credentialImage,
+    isValidating: isImageValidating,
+    error,
+  } = useSWRImmutable(
+    shouldFetchImage
       ? `/assets/credentials/image?guildId=${id}&guildAction=${credentialType}`
       : null
   )
@@ -65,6 +120,9 @@ const MintCredentialProviderComponent = ({
         ...guildCheckoutContext,
         credentialType,
         credentialImage,
+        isImageValidating,
+        isInvalidImage,
+        isTooSmallImage,
         error,
         mintedTokenId,
         setMintedTokenId,
